livekit-agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-google
livekit-plugins-bey
python-dotenv ~= 1.0
